Place pie and bar chart in their section
