+++ src/components/Layout/ChartSection.story.jsx
@@ -1,0 +1,10 @@
+import React from 'react'
+
+import * as Charts from './ChartSection'
+
+export default {
+	component: Charts.ChartSection,
+	title: 'Components/Data Visualization',
+}
+
+export const ChartSection = () => <Charts.ChartSection />
