+++ src/components/Layout/TableChartSection.jsx
@@ -2,7 +2,7 @@ import { Card } from '@blueprintjs/core'
 import { styled } from 'linaria/react'
 import React from 'react'
 
-import { PieChart } from '../Chart/PieChart'
+import { ChartSection } from './ChartSection'
 import { TableSection } from './TableSection'
 
 const StyledTableChartsSection = styled.section`
@@ -11,24 +11,8 @@ const StyledTableChartsSection = styled.section`
 	height: calc(100vh - 3.643em);
 `
 
-const Chart = styled(Card)`
-	height: 380px;
-	margin-bottom: 20px;
-`
-
 const S = {
 	TableChartSection: StyledTableChartsSection,
-	Chart: Chart,
-}
-
-export const ChartSection = () => {
-	return (
-		<section id="Chart">
-			<S.Chart>
-				<PieChart />
-			</S.Chart>
-		</section>
-	)
 }
 
 export const TableChartSection = () => {
